remove SHOPT_REGRESS (re: 506bd2b2); tweak build system
SHOPT_REGRESS is the ksh compile-time option that enables
instrumented intercepts of system calls and the like. It was an
experiment that was quietly added in 93t+ 2009-04-15 and then
apparently forgotten about, as no change was ever made after that.

By 93u+ 2012-08-01 (which we forked), enabling it was crashing the
shell, which was fixed in commit 506bd2b2, but it still doesn't
really work consistently: in my experiments with the associated
ksh-regress.rt regression tests, I noticed that the output varies
based on (a) the operating system, and (b) the presence of profile
scripts such as /etc/ksh.kshrc on the system. No one has time to
look into this further, so this commit removes that dead weight.

____
This commit also makes related tweaks to the build system:

The statically linked shell was crashing after removing regress.c,
because regress.o with its intercepts was still present in
libshell.a. To avoid similar in future, the code to link a static
library (currently repeated in five Mamfiles) is consolidated into
a new MAM include file, src/cmd/INIT/include/link_ar.mam, and
extended with code to remove object files whose sources are no
longer part of the code base. In addition, the 'ar' command usage
is updated to conform with the POSIX standard (options start with
a '-'), and the obsolete `ranlib` invocation is removed.

In mamake, the 'incl' command (added in 8ef5933d) no longer
saves/clears and restores the automatic variables, as we need them
to work normally for link_ar.mam. They are now once again saved and
restored only when the 'bind' command includes a file.