[v1.1] add &>>foo as a shorthand for >>foo 2>&1 (re: 921bbcae)
This is the 'append' variant of &>. It is very simple to add with
a little code in the lexer (lex.c); all we need to do is xor SYMREP
into the token to signify that the > is repeated. The parser will
then do the right thing without modification.

&>> msut be disabled in the POSIX compatibility mode for the same
reason as &> because, POSIXly,
	somecommand &>>somefile
is the same as
	somecommand &
	>>somefile
so it's a syntactical incompatibility. It's not one that's likely
to cause problems in real-life scripts, but better safe than sorry.