arith: do not pointlessly reinitialise Inf and NaN
There are special global static C variables for the float values
Inf and NaN (i.e. Inf, Infnod, NaN, NaNnod). These are initialised
with a strtold() call every time an arith expression references Inf
or NaN, but that is not necessary because, though the result of
that call is system-dependent, it is constant. So only initialise
them on first use.