Throw error on superflous operands in test/[
In https://bugs.debian.org/1068340, Vincent Lefevre reports:

>   $ test a = a b ; echo $?
>   0
>
> while this should be an error.

Compound expressions with -a or -o or parentheses (deprecated by
POSIX) do throw the expected error:

  $ test 1 -eq 1 -a a = a b ; echo $?
  -ksh: test: incorrect syntax
  2
  $ test \( a = a b \); echo $?
  -ksh: test: incorrect syntax
  2

The behaviour is identical in the original Bourne shell (tested
/bin/sh on Solaris 9) and on ksh88. However, every other shell
throws an error, and so do external 'test' commands on macOS and
Linux.

Fixing the test/[ mess without unwanted side effects is fiendishly
difficult, but might be made easier by the fact that this bug only
affects simple expressions. This commit tries to fix it by adding a
distinct function that only handles the specific case of a single
unary or binary expression. It's a kludge, but it avoids messing
with existing code to minimise the risk of side effects.

src/cmd/ksh93/bltins/test.c:
- Add check_toomanyops() function. It assumes any initial '!' is
  skipped. It checks the argument list for a simple binary or unary
  expression, then checks if it is followed by another argument
  that is not '-a' or '-o' (which is marked by TEST_ANDOR); it
  throws a "too many operands" error if found.
- When abandoning the final ] matching [, actually set that
  argument to NULL, properly terminating the arguments list. The
  new function needs this and it does not hurt anything else.
- Call the new function after processing a possible '!' negator,
  passing the arguments list with any '!' skipped.

Resolves: https://github.com/ksh93/ksh/issues/739