ksh, libcmd: remove EBCDIC support
EBCDIC is the basic character set on IBM mainframes that run z/OS
or its prececessor MVS. Everything else uses ASCII as its lowest
common denominator. As I have not succeeded in finding any way to
get access to an IBM mainframe to test ksh on z/OS, the EBCDIC code
is now removed as it has surely fallen victim to bit rot by now (if
it hadn't already in the 2012 version that we forked).