libast: remove unused/obsolete file magic and MIME type libraries
The file magic library and database was used for the AT&T
implementation of the 'file' command and (per the ast-open-archive
repo) was also used by 'tw', 'mailx', and 'nmake'. We don't ship
any of them, so it's unused here. It's also been unmaintained since
2012 and was out of date even then -- for instance, Microsoft's
ubiquitous OOXML format (.docx, etc.) is not included in magic.tab,
though it was introduced in 2007.

The MIME type library, which uses the magic library and database,
is similarly out of date. It was only used by 'mailx'.