Build system tweaks; other small tweaks & cleanups
Main changes:

libast: Reduce the dependencies on ast_release.h (which defines
_AST_release for release builds, disabling debugging code). No
longer auto-include this from ast.h. Files that need it should
include it explicitly and declare this dependency in the Mamfile.

bin/package: Rework the way of dealing with LD_LIBRARY_PATH and
friends. On init, ensure $INSTALLROOT-based directories are *not*
in there (potentially inherited from a 'bin/package use'
environmetn) as this may break the build. When doing 'test' and/or
'use', re-add these as appropriate.