GitHub CI: don't run tests with unsupported locales (re: 09d8be78)
Now that ksh initialises the locale correctly, the nl_NL.UTF-8
tests are failing on the CI with 'unsupported locale' messages. An
experiment on a test branch revealed that the GitHub CI runners
don't have support for non-USA locales installed at all. The output
of 'locale -a' on them is very short:

    C
    C.utf8
    POSIX
    en_US.utf8

.github/workflows/ci.yml:
- Replace the invalid nl_NL.UTF-8 and ja_JP.SJIS test runs with an
  en_US.utf8 run.