shtests: do not run locale.sh with shcomp(1) by default
In the ja_JP.SJIS locale, running the locale.sh tests under shcomp
triggers a syntax error due to literal UTF-8 strings in the test
scripts. Reproducer on macOS:

 $ LANG=ja_JP.SJIS bin/shtests -lc locale
 (result: syntax error in locale.sh)

We could work around that with trickery, but it's not worth it,
because running locale.sh under shcomp doesn't really make sense.
The locale.sh tests change the locale on the fly, and as the shell
parser is also affected by the locale, the parsing of each line in
the script depends on the execution of prior lines in certain
contexts. This should be expected to fail with shcomp, as the file
is parsed entirely before executing.

src/cmd/ksh93/tests/shtests:
- Like pty.sh, do not run locale.sh under shcomp, unless -c or
  --compile is explicitly specified.