Fix UB with subtracting from or adding to null pointers
Various points in the code use a strange trick to typecast a
pointer to an integer: they subtract a NULL pointer from the
pointer, counting on getting an integer typecast of the address as
a result. To convert an integer into a pointer, a NULL pointer is
added. But the C standard is clear: in both cases, it's undefined
behaviour. This commit changes that all into proper typecasts.

Further reading:
https://wiki.sei.cmu.edu/confluence/x/1dYxBQ
https://wiki.sei.cmu.edu/confluence/x/0dUxBQ