simplify ${.sh.version} / $((.sh.version)) init
We don't need to determine the length of the e_version string at
runtime. Since e_version is declared as a static array, its size is
known at compile time, and the string length is a function of that,
so this is now done at compile time to save cycles on init.

This commit also removes a few tests for library functions that we
can now safely assume to be present: nice(3) and fork(2), as well
as pathnative() and pathposix() which are provided by libast
(unlike the old AT&T practice, we always compile ksh against our
shipped and known version of libast).