edit.c: fix out of bounds write in output buffer
When allocating the output buffer, the pointer to the last byte,
ep->e_outlast, is set to one past the end of the buffer. This can
cause an out of bounds write in ed_putbyte()/ed_putchar() while
setting the terminating zero byte. Fix this by setting it to the
last byte of the buffer instead.